Update dataset-doc-template.md
Adjusted capitalization and added id's as a standard to each section. Not sure if we'll need id's but it will be better to have this template with as much detail as possible so we don't need to go rework a lot of docs in the future if we decide to add later.  Other than that, looks great!